+++ NEWS.md
@@ -6,12 +6,57 @@ since the **3.0.0** release, except for bug fixes.
 Note that each entry is kept to a minimum, see links for details.
 
 ## Language changes
+
 * Pin operator now takes an expression. [[Feature #17411]]
 
     ```ruby
     Prime.each_cons(2).lazy.find_all{_1 in [n, ^(n + 2)]}.take(3).to_a
     #=> [[3, 5], [5, 7], [11, 13]]
     ```
+
+* Multiple assignment evaluation order has been made consistent with
+  single assignment evaluation order.  With single assignment, Ruby
+  uses a left-to-right evaluation order.  With this code:
+
+    ```ruby
+    foo[0] = bar
+    ```
+
+  The following evaluation order is used:
+
+  1. `foo`
+  2. `bar`
+  3. `[]=` called on the result of `foo`
+
+  In Ruby before 3.1.0, multiple assignment did not follow this
+  evaluation order.  With this code:
+
+    ```ruby
+    foo[0], bar.baz = a, b
+    ```
+
+  Versions of Ruby before 3.1.0 would evaluate in the following
+  order
+
+  1. `a`
+  2. `b`
+  3. `foo`
+  4. `[]=` called on the result of `foo`
+  5. `bar`
+  6. `baz=` called on the result of `bar`
+
+  Starting in Ruby 3.1.0, evaluation order is now consistent with
+  single assignment, with the left hand side being evaluated before
+  the right hand side:
+
+  1. `foo`
+  2. `bar`
+  3. `a`
+  4. `b`
+  5. `[]=` called on the result of `foo`
+  6. `baz=` called on the result of `bar`
+
+  [[Bug #4443]]
 
 ## Command line options
 
@@ -96,6 +141,7 @@ Excluding feature bug fixes.
 ## Miscellaneous changes
 
 
+[Bug #4443]: https://bugs.ruby-lang.org/issues/4443
 [Feature #12194]: https://bugs.ruby-lang.org/issues/12194
 [Feature #14256]: https://bugs.ruby-lang.org/issues/14256
 [Feature #15198]: https://bugs.ruby-lang.org/issues/15198